+++ src/components/Header/header.tsx
@@ -1,12 +1,10 @@
+import { useEffect } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 
 import { useAppDispatch, useAppSelector } from 'redux/hooks';
-import { logOut } from 'redux/authSlice';
+import { logOut, setOpenUserPage } from 'redux/authSlice';
 
-import { useEffect } from 'react';
 import { HeaderUserButtons, HeaderUserLinks } from './HeaderButtons';
-
-import { setOpenUserPage } from 'redux/authSlice';
 
 import styles from './header.module.scss';
