+++ src/pages/TemplatePage/templatePage.tsx
@@ -1,8 +1,7 @@
 import { Suspense } from 'react';
 import { Outlet } from 'react-router-dom';
-import Loader from 'components/Loader';
-import Header from 'components/Header/header';
-import Footer from 'components/Footer/footer';
+
+import { Header, Footer, Loader } from 'components';
 
 import './templatePage.scss';
