+++ src/routes/privateRoute.tsx
@@ -1,4 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
+
 import { useAppSelector } from 'redux/hooks';
 
 const PrivateRoute = () => {
